Extract interface for TaskExecutor
In preparation for adding a ThreadPerSplitTaskExecutor
