+++ core/trino-main/src/test/java/io/trino/execution/executor/timesharing/Histogram.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.execution.executor;
+package io.trino.execution.executor.timesharing;
 
 import com.google.common.collect.ImmutableList;
